### 2020-05-24 1ca1ce1
Add missing DOIs in paper references
