+++ paper/paper.bib
@@ -14,7 +14,8 @@ Photonic crystals
   number={22},
   pages={12576--12578},
   year={2003},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={https://doi.org/10.1073/pnas.2133313100}
 }
 @article{yablonovitch1987inhibited,
   title={Inhibited spontaneous emission in solid-state physics and electronics},
@@ -24,7 +25,8 @@ Photonic crystals
   number={20},
   pages={2059},
   year={1987},
-  publisher={APS}
+  publisher={APS},
+  doi={https://doi.org/10.1007/978-1-4615-1963-8_41}
 }
 @article{john1987strong,
   title={Strong localization of photons in certain disordered dielectric superlattices},
@@ -34,7 +36,8 @@ Photonic crystals
   number={23},
   pages={2486},
   year={1987},
-  publisher={APS}
+  publisher={APS},
+  doi={https://doi.org/10.1103/physrevlett.58.2486}
 }
 @article{joannopoulos2008molding,
   title={Molding the flow of light},
@@ -50,7 +53,8 @@ Photonic crystals
   number={6950},
   pages={847--851},
   year={2003},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={https://doi.org/10.1038/nature01940}
 }
 
 
@@ -83,7 +87,8 @@ Electronic TIs
   number={5806},
   pages={1757--1761},
   year={2006},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={https://doi.org/10.1126/science.1133734}
 }
 
 
@@ -96,7 +101,8 @@ Photonic TIs
   number={22},
   pages={223901},
   year={2015},
-  publisher={APS}
+  publisher={APS},
+  doi={https://doi.org/10.1103/physrevlett.114.223901}
 }
 @article{blanco2019engineering,
   title={Engineering fragile topology in photonic crystals: Topological quantum chemistry of light},
@@ -106,7 +112,8 @@ Photonic TIs
   number={3},
   pages={032005},
   year={2019},
-  publisher={APS}
+  publisher={APS},
+  doi={https://doi.org/10.1103/physrevresearch.1.032005}
 }
 @article{blanco2020tutorial,
   title={Tutorial: computing topological invariants in 2D photonic crystals},
@@ -116,7 +123,8 @@ Photonic TIs
   number={2},
   pages={1900117},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1002/qute.201900117}
 }
 @article{rider2019perspective,
   title={A perspective on topological nanophotonics: Current status and future challenges},
@@ -126,5 +134,6 @@ Photonic TIs
   number={12},
   pages={120901},
   year={2019},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi={https://doi.org/10.1063/1.5086433}
 }
