### 2020-05-24 0791c61
Fix some capitalisation etc in the paper references
