+++ paper/paper.bib
@@ -42,13 +42,13 @@ Photonic crystals
 @article{joannopoulos2008molding,
   title={Molding the flow of light},
   author={Joannopoulos, John D and Johnson, Steven G and Winn, Joshua N and Meade, Robert D},
-  journal={Princeton Univ. Press, Princeton, NJ [ua]},
+  journal={Princeton Univ. Press, Princeton, NJ},
   year={2008}
 }
 @article{knight2003photonic,
   title={Photonic crystal fibres},
   author={Knight, Jonathan C},
-  journal={nature},
+  journal={Nature},
   volume={424},
   number={6950},
   pages={847--851},
@@ -60,7 +60,7 @@ Photonic crystals
 
 Electronic TIs
 @article{kane2005z,
-  title={Z 2 topological order and the quantum spin Hall effect},
+  title={$Z_{2}$ topological order and the quantum spin Hall effect},
   author={Kane, Charles L and Mele, Eugene J},
   journal={Physical review letters},
   volume={95},
@@ -82,7 +82,7 @@ Electronic TIs
 @article{bernevig2006quantum,
   title={Quantum spin Hall effect and topological phase transition in HgTe quantum wells},
   author={Bernevig, B Andrei and Hughes, Taylor L and Zhang, Shou-Cheng},
-  journal={science},
+  journal={Science},
   volume={314},
   number={5806},
   pages={1757--1761},
