### 2020-05-24 #2239 @sp94
Thanks for adding me to the waiting list!

From the list of possible reviewers, here are some in Physics/optics/solid state/condensed matter theory, and who also lists Julia in their programming languages:
- crstnbr
- mdavezac
- frssp
- solanpaa
