### 2020-06-06 #2239 @kyleniemeyer
Hi @jgostick, can you take on this submission as your first editorial assignment?
