### 2020-06-06 #2239 @jgostick
@kyleniemeyer Does it matter that I don't know what a photonic crystal is, have not yet started to use Julia, and can't think of single person who could review this?
