### 2020-06-08 d738028
Bug fix: spacing of k-points on band diagram was defaulting to zero, leading to divide by zero error
