+++ src/band_diagrams.jl
@@ -87,7 +87,7 @@ end
 Plot the bands generated by `solve(solver, k, polarisation)` along `ks`.
 """
 function plot_band_diagram(solver::Solver, ks, polarisation::Polarisation;
-            dk=0, labels=[], bands=1:10, frequency_scale=1, color="k", markersize=nothing)
+            dk=nothing, labels=[], bands=1:10, frequency_scale=1, color="k", markersize=nothing)
     # Convert BrillouinZoneCoordinate to labelled positions in k space
     if labels == []
         labels = [hasproperty(x,:label) ? x.label : "" for x in ks]
