### 2020-06-08 eb67b29
Bug fix: The Zoo had the wrong permittivities for Zhu crystal
