### 2020-06-08 54ea4be
Remove default cutoff from make_wu_generic
