### 2020-06-12 #2239 @dpsanders
👋 Hi @sp94! Thanks for submitting to JOSS and sorry for the delay.

This looks like a very interesting submission. The first thing I notice browsing through is that there are basically no docs. Could you please write some narrative and API documentation? You can use e.g. Documenter.jl for this.

Also the paper is quite short compared to most JOSS papers. Could you please review the requirements, such as including comparisons to other software and stating the research need?

Thanks!
