### 2020-06-22 #2239 @dpsanders
Hi @sp94! Sorry, I've been out of action for a while but am now picking up again.

Could you please add a couple of examples to the paper too? And briefly comment what a photonic crystal is.

Are photonic crystals in 3D also relevant? How hard would it be to extend the package to that case? Is that in the pipeline?

For the documentation aspect, [Documenter.jl](https://github.com/JuliaDocs/Documenter.jl) can automatically harvest the docstrings and put them together to make the API docs. I recommend that you expand what you have in the README into the narrative docs. As a useful guide to the different types of documentation I recommend https://documentation.divio.com/.

Thanks!
