### 2020-06-26 cb0d1a3
Add travis file to automatically deploy documentation
