### 2020-06-26 13eb27e
Initial commit of make file for Documenter
