+++ docs/make.jl
@@ -1,0 +1,6 @@
+using Documenter, Peacock
+
+makedocs(sitename="Peacock.jl Documentation")
+
+deploydocs(
+	repo="github.com/sp94/Peacock.jl.git")
