### 2020-06-26 1ca81bc
Include Peacock.jl in the docs build environment
