+++ docs/Project.toml
@@ -1,2 +1,3 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Peacock = "b643ced8-3934-11ea-3f5f-690af0bb8ffb"
