### 2020-06-26 04f6511
Install matplotlib in docs build environment
