### 2020-06-27 22a8cd8
Add docs badge to README
