+++ README.md
@@ -2,6 +2,8 @@
 <img src="logo.png" alt="Peacock logo"></img>
 </div>
 
+[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://sp94.github.io/Peacock.jl/stable)
+
 ## Introduction
 
 `Peacock.jl` - or the **P**lane-wave **E**xpansion **A**pproach to **C**haracterising **O**ptical **C**rystals in **k**-space - is a Julia package for studying photonic crystals using the Plane Wave Expansion Method.
