### 2020-06-27 348d0f8
Move docs logo into assets folder
