+++ docs/src/index.md
@@ -1,4 +1,4 @@
-![](logo.png)
+![](assets/logo.png)
 
 # Introduction
 
