### 2020-06-27 93f5c91
Move peacock_feathers_zoom.png to asset folder
