### 2020-06-27 c984f3c
Remove duplicate of logo and zoom images
