+++ docs/src/reference/index.md
@@ -1,0 +1,4 @@
+# Technical reference
+```@autodocs
+Modules = [Peacock]
+```
