+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#","page":"Introduction","title":"Introduction","text":"(Image: )","category":"page"},{"location":"#Introduction-1","page":"Introduction","title":"Introduction","text":"","category":"section"},{"location":"#","page":"Introduction","title":"Introduction","text":"Peacock.jl - or the Plane-wave Expansion Approach to Characterising Optical Crystals in k-space - is a Julia package for studying photonic crystals using the Plane Wave Expansion Method.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"Solve for...\nTransverse electric (TE) and transverse magnetic (TM) modes of 2D photonic crystals\nNon-orthogonal unit cells\nInhomogeneous permittivity and/or permeability\nChern numbers of topological photonic crystals using built-in Wilson loop methods","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"Focused on ease of use\nInstall with one line in Julia's package manager\nSimple visualisation of geometry, fields, and fully labelled band diagrams","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"Photonic crystals are materials whose optical properties arise from the structuring of the material when the size of the structures are comparable to the wavelengths of light. Peacock.jl is named for the irridescent colours of peacock feathers which arise not from pigmentation but from their photonic crystal structure, as shown below.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"(Image: Different zooms of a Peacock) Peacock by allanlau2000 from pixabay. Feather by suju from pixabay. Electron microscope image of photonic crystal structure from Zi, Jian, et al. \"Coloration strategies in peacock feathers.\",  Proceedings of the National Academy of Sciences 100.22 (2003): 12576-12578. Copyright (2003) National Academy of Sciences.","category":"page"},{"location":"#","page":"Introduction","title":"Introduction","text":"As well as occuring naturally as in animals such as peacocks, advances in nanofabrication mean that 'designer' photonic crystals can now be manufactured for unprecedented control over the flow of light, with applications ranging from optical fibers to photonic circuitry. Photonic crystals are also a promising platform for more exotic materials like topological insulators.","category":"page"},{"location":"#Index-1","page":"Introduction","title":"Index","text":"","category":"section"},{"location":"#","page":"Introduction","title":"Introduction","text":"Pages = [\n        \"tutorials/getting_started.md\",\n        \"reference/index.md\"]","category":"page"}]
[…]
