### 2020-06-27 45c4224
Add horizontal line between logo and docs badge
