### 2020-06-27 db8febd
Add a proper logo, rename old logo to banner
