+++ docs/src/index.md
@@ -1,4 +1,4 @@
-![](assets/logo.png)
+![](assets/banner.png)
 
 # Introduction
 
