### 2020-06-27 5e49b00
Add ‘background’, ‘features’, ‘limitations’, ‘contributing’ to landing page of docs
