### 2020-06-27 e2b183c
Initial commit of ‘how to load crystal from zoo’ guide
