### 2020-06-27 bfc7b61
Split technical reference into public and private
