+++ docs/src/reference/index.md
@@ -1,4 +1,17 @@
 # Technical reference
+
+## Public
+
 ```@autodocs
 Modules = [Peacock]
+Private = false
+```
+
+## Private
+
+!!! note "The `Peacock.jl` namespace" The following objects are used internally, and aren't exported to the global namespace. However, they may be useful for more advanced users, in which case you should use `Peacock.[name]` to access these.
+
+```@autodocs
+Modules = [Peacock]
+Public = false
 ```
