+++ docs/src/tutorials/getting_started.md
@@ -1,0 +1,179 @@
+```@setup tutorial
+using Peacock
+using PyPlot
+```
+# Getting started
+
+This tutorial will help you get started with `Peacock.jl` for the first time.
+
+We will install `Peacock.jl`, and then use it to study an example photonic crystal from a book (Joannopoulos *et al* 2008) by plotting the band structure and visualising the modes of the crystal.
+
+## Installation
+
+Before using `Peacock.jl` for the first time, you should install it using the built-in Julia package manager.
+```julia
+using Pkg
+Pkg.add("Peacock")
+```
+
+After installation, the `Peacock.jl` can be loaded in Julia.
+```julia
+using Peacock
+```
+
+For this tutorial, we'll also install and load `PyPlot` to control our figures.
+```julia
+using Pkg
+Pkg.add("PyPlot")
+using PyPlot
+```
+
+
+## Defining the photonic crystal
+
+In this section we will create a  [`Geometry`](@ref) that holds the shape, size, and materials of the crystal.
+
+The crystal we will reproduce in this tutorial is taken from chapter 5 of Joannopoulos *et al* 2008. It consists of dielectric cylinders (``\epsilon_\mathrm{cyl}=8.9, \mu_\mathrm{cyl}=1``) in air (``\epsilon_\mathrm{air}=1, \mu_\mathrm{air}=1``). The cylinders are arranged on a square lattice with separation ``a``, and each cylinder has a radius of ``r=a/5``.
+
+First, let's define the functions `epf(x,y)` and `muf(x,y)`, which return the permittivity and permeability of the unit cell at ``(x,y)``, where ``(0,0)`` is the center of the unit cell. We will work in units of length where the separation between cylinders is unity, ``a=1``, such the radius of each cylinder is `0.2`.
+```@example tutorial
+# Permittivity
+function epf(x,y)
+	# equation of a circle with radius 0.2a
+    if x^2+y^2 <= 0.2^2
+    	# dielectric inside the circle
+        return 8.9
+    else
+    	# air outside the circle
+        return 1
+    end
+end
+
+# Permeability is unity everywhere
+function muf(x,y)
+    return 1
+end
+
+nothing # hide
+```
+
+Now we declare the lattice parameters. The cylinders are on a square lattice, so our lattice vectors are orthogonal and of equal length `a`.
+```@example tutorial
+a1 = [1, 0]  # first lattice vector
+a2 = [0, 1]  # second lattice vector
+nothing # hide
+```
+
+We must also give the resolution at which the geometry should be generated at.
+```@example tutorial
+d1 = 0.01  # resolution along first lattice vector
+d2 = 0.01  # resolution along second lattice vector
+nothing # hide
+```
+A smaller value of `d1` or `d1` will result in a higher resolution grid.
+
+Finally, we are ready to construct and visualise our [`Geometry`](@ref).
+```@example tutorial
+geometry = Geometry(epf, muf, a1, a2, d1, d2)
+plot(geometry)
+
+figure(1); savefig("ep.png") # hide
+figure(2); savefig("mu.png") # hide
+nothing # hide
+```
+
+![](ep.png)
+
+![](mu.png)
+
+
+## Setting up the solver
+
+In this section we will create a [`Solver`](@ref) that approximates the geometry using a truncated Plane Wave Expansion (see further reading). The number of plane waves is determined by the cutoff. Increasing the cutoff will increase the accuracy of the solution, but low-contrast photonic crystals can be well approximated with a relatively small basis of plane waves.
+```@example tutorial
+fourier_space_cutoff = 7
[…]
