### 2020-06-27 14cbbc9
Give docs an explicit sidebar layout
