### 2020-06-27 e058b28
Add more dependencies to the docs build environment
