### 2020-06-27 98b5a5c
Ensure travis activates development environment
