+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"reference/#Technical-reference-1","page":"Technical reference","title":"Technical reference","text":"","category":"section"},{"location":"reference/#","page":"Technical reference","title":"Technical reference","text":"## Public","category":"page"},{"location":"reference/#","page":"Technical reference","title":"Technical reference","text":"Modules = [Peacock]\nPrivate = false","category":"page"},{"location":"reference/#Peacock.BrillouinZoneCoordinate","page":"Technical reference","title":"Peacock.BrillouinZoneCoordinate","text":"BrillouinZoneCoordinate(p::Float64, q::Float64, label::String=\"\")\n\nA labelled coordinate in the Brillouin zone.\n\nThe arguments p and q are the coefficients of reciprocal lattice vectors b1 and b2. The k-space coordinate, k = p * b1 + q * b2, is generated by get_k(coord::BrillouinZoneCoordinate, basis::PlaneWaveBasis). For example, BrillouinZoneCoordinate(0.5,0) is on the edge of the first Brillouin zone.\n\n\n\n\n\n","category":"type"},{"location":"reference/#Peacock.Geometry","page":"Technical reference","title":"Peacock.Geometry","text":"2D geometry defined in real space.\n\n\n\n\n\n","category":"type"},{"location":"reference/#Peacock.Geometry-Tuple{Function,Function,Array{#s18,1} where #s18<:Real,Array{#s19,1} where #s19<:Real,Real,Real}","page":"Technical reference","title":"Peacock.Geometry","text":"Geometry(epf::Function, muf::Function, a1::Array{<:Real,1}, a2::Array{<:Real,1}, d1::Real, d2::Real)\n\nGenerate geometry with permittivity epf(x,y) and permeability muf(x,y).\n\nThe real space lattice vectors, a1 and a2, define the unit cell. The grid resolution along each lattice vector is d1 and d2, respectively.\n\n\n\n\n\n","category":"method"},{"location":"reference/#Peacock.Geometry-Tuple{Function,Function,Real,Real,Real,Real}","page":"Technical reference","title":"Peacock.Geometry","text":"Geometry(epf::Function, muf::Function, a1::Array{<:Real,1}, a2::Array{<:Real,1}, d1::Real, d2::Real)\n\nGenerate geometry with permittivity epf(x,y) and permeability muf(x,y).\n\nThe real space lattice vectors are assumed to have unit length and are at angles a1_deg and a2_deg counter-clockwise from the x-axis. The grid resolution along each lattice vector is d1 and d2, respectively.\n\n\n\n\n\n","category":"method"},{"location":"reference/#Peacock.HilbertSpace-Tuple{Array{Mode,1}}","page":"Technical reference","title":"Peacock.HilbertSpace","text":"HilbertSpace(modes::Array{Mode,1})\n\nReturns the Hilbert space spanned by the modes.\n\nThe data of the Hilbert space is guaranteed to be orthonormal under the weighting of the modes.\n\n\n\n\n\n","category":"method"},{"location":"reference/#Peacock.Mode","page":"Technical reference","title":"Peacock.Mode","text":"Mode(k0, frequency, data, weighting, basis, label)\n\nEigenmode of a photonic crystal expressed on a plane-wave basis with a weighted inner product.\n\n\n\n\n\n","category":"type"},{"location":"reference/#Peacock.Solver-Tuple{Geometry,Int64}","page":"Technical reference","title":"Peacock.Solver","text":"Solver(geometry::Geometry, cutoff::Int)\n\nApproximate the geometry using a truncated bas
[…]
