### 2020-06-27 55b26b5
Switch back to pre-rendered figures rather than generating with the docs
I was having trouble with figures rendering in low quality, and also the PyPlot axis getting mixed up between different files
