+++ docs/src/how-tos/zoo.md
@@ -1,12 +1,3 @@
-```@setup zoo
-using PyPlot
-using Parameters
-
-using Peacock
-using Peacock.Zoo
-fourier_space_cutoff = 7
-```
-
 # How to load a crystal from the Zoo
 
 ## Namespaces
@@ -17,20 +8,15 @@ The `Peacock.Zoo` module contains functions to generate some of the photonic cry
 ## Unpacking parameters
 
 Each `make_*` function returns a `NamedTuple` containing the geometry, solver, polarisation, and high symmetry ``k``-points of the crystal. For example, the topological photonic crystal first introduced by Wu *et al* 2015 can be generated using `make_wu_topo`.
-```@example zoo
+```julia
 # Load photonic topological insulator from Wu et al 2015
 @unpack geometry, solver, polarisation, G, K, M = make_wu_topo(fourier_space_cutoff)
 
 # Preview the crystal
 plot(geometry)
-
-figure(1); savefig("wu_ep.png") # hide
-figure(2); savefig("wu_mu.png") # hide
 ```
 
-![](wu_ep.png)
-
-![](wu_mu.png)
+![](../figures/example_zoo_geometry.png)
 
 
 ## Available crystals
