### 2020-06-28 aa7cfdc
Fix headers that had “non-breaking space” instead of regular space, and add some id’s to headers
