### 2020-06-28 c9f85f0
Add “how-to guide” on topology and Wilson loops
