### 2020-06-28 7221ef7
Remove index from docs landing page
