### 2020-06-28 7354587
Expand the Wilson loops how-to guide
