### 2020-06-28 3b18dd7
Cutdown the README and provide links to the Docs instead
