### 2020-06-28 ca0fd1d
Replace simple LaTeX expression with ASCII
