### 2020-06-28 #2239 @sp94
Hi @dpsanders,

I just want to update you with where I'm at.

Thanks for recommending Documenter.jl and the divio guide to documentation, they were very helpful. I've ported the old README over to the docs, and split it up into intro/tutorial/how-to's/reference. There are a couple of docstrings still to fix in the reference, and I need to write the "contributing" statement, but I think everything else is ready.

> Also the paper is quite short compared to most JOSS papers. Could you please review the requirements, such as including comparisons to other software and stating the research need?
> Could you please add a couple of examples to the paper too? And briefly comment what a photonic crystal is.
OK, I'll be working on this today/tomorrow.

> Are photonic crystals in 3D also relevant? How hard would it be to extend the package to that case? Is that in the pipeline?
3D is relevant and important, so I have added "2D only" to the README. However, 2D can still be useful for a first approximation of a 3D crystal, and there is still more to learn about 2D crystals with "topology". The method itself can be extended quite easily to 3D, really I think the difficulty is how to keep the interface simple for both 2D and 3D. I will open a GitHub issue with some options for this.

I will let you know in the next few days after I have
(1) completed the few remaining docstrings
(2) reviewed the paper requirements, and added examples to the paper
(3) opened a GitHub issue for extending to 3D
