### 2020-06-30 eedcec3
Add contributions guidelines
