### 2020-06-30 0704c95
Improve description of convergence limitation for high contrast materials
