### 2020-06-30 72b8b93
Add hyperlinks to contributions page
