### 2020-06-30 f4c982a
Redirect paper zoom figure to the docs source
