### 2020-06-30 12562a8
Add example figures to paper
