+++ paper/paper.md
@@ -64,6 +64,8 @@ insulator [@kane2005z; @kane2005quantum; @bernevig2006quantum] for which the
 to identify topological photonic crystals using Wilson loops
 [@blanco2020tutorial], for which open source implementations have been lacking.
 
+![Examples. (a) ... (b) ..... \label{fig:examples}](examples.svg)
+
 # Acknowledgements
 
 S.J.P. acknowledges his studentship from the Centre for Doctoral Training on
