### 2020-06-30 50904a0
Paper not compiling, try replacing examples svg with pdf
