### 2020-07-04 f0cb13d
paper: caption the examples figure
