### 2020-07-04 db40e93
paper: focus on research need and comparison to other software
