### 2020-07-04 fa25b40
paper: add citation to 'Band topology in classical waves...' by Wang et al 2019
