+++ paper/paper.bib
@@ -137,3 +137,13 @@ Photonic TIs
   publisher={AIP Publishing LLC},
   doi={https://doi.org/10.1063/1.5086433}
 }
+@article{wang2019band,
+  title={Band topology in classical waves: Wilson-loop approach to topological numbers and fragile topology},
+  author={Wang, Hai-Xiao and Guo, Guang-Yu and Jiang, Jian-Hua},
+  journal={New Journal of Physics},
+  volume={21},
+  number={9},
+  pages={093029},
+  year={2019},
+  publisher={IOP Publishing}
+}
