### 2020-07-04 4ae5c89
paper: add explicit comparison to other software
