+++ paper/paper.bib
@@ -147,3 +147,25 @@ Photonic TIs
   year={2019},
   publisher={IOP Publishing}
 }
+
+Other software
+@article{johnson2001block,
+  title={Block-iterative frequency-domain methods for Maxwell's equations in a planewave basis},
+  author={Johnson, Steven G and Joannopoulos, John D},
+  journal={Optics express},
+  volume={8},
+  number={3},
+  pages={173--190},
+  year={2001},
+  publisher={Optical Society of America}
+}
+@article{gresch2017z2pack,
+  title={Z2Pack: Numerical implementation of hybrid Wannier centers for identifying topological materials},
+  author={Gresch, Dominik and Autes, Gabriel and Yazyev, Oleg V and Troyer, Matthias and Vanderbilt, David and Bernevig, B Andrei and Soluyanov, Alexey A},
+  journal={Physical Review B},
+  volume={95},
+  number={7},
+  pages={075146},
+  year={2017},
+  publisher={APS}
+}
