### 2020-07-04 1a72d6b
paper: remove accidental bracket
