### 2020-07-04 45a03cd
paper: fix grammar in example caption
