### 2020-07-04 #2239 @sp94
Hi @dpsanders, please let me know if the paper addresses your comments about adding examples, research need, and comparisons to other software
