+++ src/Zoo.jl
@@ -43,11 +43,18 @@ function wu_ep(x, y, angles, Rs, d1s, d2s, ep_bg, ep_cyl)
     return ep_bg
 end
 
-function make_wu_generic(Rs, d1s, d2s, polarisation, P; angles=[60n for n in 0:5], ep_bg=1, ep_cyl=11.7)
+"""
+    make_wu_generic(Rs, d1s, d2s, polarisation, cutoff;
+                angles=[60n for n in 0:5], ep_bg=1, ep_cyl=11.7)
+
+Returns a `NamedTuple` of `geometry`, `solver`, `polarisation`, and the `G`,
+`M`, `K` high-symmetry points for photonic crystals similar to Wu et al, 2015.
+"""
+function make_wu_generic(Rs, d1s, d2s, polarisation, cutoff; angles=[60n for n in 0:5], ep_bg=1, ep_cyl=11.7)
     epf(x,y) = wu_ep(x, y, angles, Rs, d1s, d2s, ep_bg, ep_cyl)
     muf(x,y) = 1.0
     geometry = Geometry(epf, muf, -60, +60, 1/500, 1/500)
-    solver = Solver(geometry, P)
+    solver = Solver(geometry, cutoff)
     G = BrillouinZoneCoordinate(  0,   0, "Γ")
     M = BrillouinZoneCoordinate(  0, 1/2, "M")
     K = BrillouinZoneCoordinate(1/3, 1/3, "K")
@@ -58,34 +65,58 @@ end
 
 
 """
-    make_wu(a0_div_R, P)
+    make_wu(a0_div_R, cutoff)
 
 Reproduces "Scheme for Achieving a Topological Photonic Crystal
-by Using Dielectric Material", Wu et al, 2015
+by Using Dielectric Material", Wu et al, 2015.
+
+Returns a `NamedTuple` of `geometry`, `solver`, `polarisation`, and the
+`G`, `M`, `K` high-symmetry points.
 """
-function make_wu(a0_div_R, P)
+function make_wu(a0_div_R, cutoff)
     # Geometry
     R = 1 / a0_div_R # assuming a0 = 1
     d = 2R/3
-    return make_wu_generic(R, d, d, TM, P)
+    return make_wu_generic(R, d, d, TM, cutoff)
 end
 
-function make_wu_triv(P::Int)
-    return make_wu(3.125, P)
+"""
+    make_wu_topo(cutoff::Int)
+
+Reproduces the trivial crystal (a0_div_R=3.125)from  "Scheme for Achieving a
+Topological Photonic Crystal by Using Dielectric Material", Wu et al, 2015.
+
+Returns a `NamedTuple` of `geometry`, `solver`, `polarisation`, and the
+`G`, `M`, `K` high-symmetry points.
+"""
+function make_wu_triv(cutoff::Int)
+    return make_wu(3.125, cutoff)
 end
 
-function make_wu_topo(P::Int)
-    return make_wu(2.9, P)
+"""
+    make_wu_topo(cutoff::Int)
+
+Reproduces the topological crystal (`a0_div_R=2.9`) from "Scheme for Achieving a
+Topological Photonic Crystal by Using Dielectric Material", Wu et al, 2015.
+
+Returns a `NamedTuple` of `geometry`, `solver`, `polarisation`, and the
+`G`, `M`, `K` high-symmetry points.
+"""
+function make_wu_topo(cutoff::Int)
+    return make_wu(2.9, cutoff)
 end
 
 
 """
-    make_wu_primitive(P)
+    make_wu_primitive(cutoff)
 
 Reproduces the primitive unit cell of "Scheme for Achieving a
 Topological Photonic Crystal by Using Dielectric Material", Wu et al, 2015
+
[…]
