### 2020-07-04 3a2ce7b
docs: add missing docstrings
