+++ src/band_diagrams.jl
@@ -85,6 +85,9 @@ end
     plot_band_diagram(solver::Solver, ks, polarisation::Polarisation, <keyword arguments>)
 
 Plot the bands generated by `solve(solver, k, polarisation)` along `ks`.
+
+Takes the same keyword arguments as [`plot_band_diagram(my_solve::Function, ks)`](@ref),
+but here `ks` can also include [`BrillouinZoneCoordinate`](@ref)s.
 """
 function plot_band_diagram(solver::Solver, ks, polarisation::Polarisation;
             dk=nothing, labels=[], bands=1:10, frequency_scale=1, color="k", markersize=nothing)
