### 2020-07-04 5411667
bug fix: default value of dk_outer and dk_inner in plot_wilson_loop_winding was causing divide by zero errors
