### 2020-07-04 41b7b7b
add marker_size option to plot_wilson_loop_winding
