### 2020-07-04 02785e3
allow ‘label’ keyword argument when plotting modes
