+++ src/plotting.jl
@@ -5,7 +5,7 @@ import PyPlot.plot
 
 Plot the real-valued `field` on a unit cell with lattice vectors `a1` and `a2`.
 """
-function plot_field(field::Array{<:Real,2}, a1::Array{<:Real,1}, a2::Array{<:Real,1}; cmap="coolwarm", vmin=nothing, vmax=nothing)
+function plot_field(field::Array{<:Real,2}, a1::Array{<:Real,1}, a2::Array{<:Real,1}; cmap="coolwarm", vmin=nothing, vmax=nothing, label=nothing)
     ps = range(-0.5, stop=0.5, length=size(field,1)+1)
     qs = range(-0.5, stop=0.5, length=size(field,2)+1)
     xs = [p*a1[1]+q*a2[1] for p in ps, q in qs]
@@ -22,6 +22,10 @@ function plot_field(field::Array{<:Real,2}, a1::Array{<:Real,1}, a2::Array{<:Rea
     pcolormesh(xs, ys, field, cmap=cmap, vmin=vmin, vmax=vmax)
     gca().set_aspect("equal")
     axis("off")
+    colorbar()
+    if label != nothing
+        title(label)
+    end
 end
 
 """
@@ -29,7 +33,7 @@ end
 
 Plot the complex-valued `field` on a unit cell with lattice vectors `a1` and `a2`.
 """
-function plot_field(field::Array{<:Complex,2}, a1::Array{<:Real,1}, a2::Array{<:Real,1}; cmap="coolwarm", vmin=nothing, vmax=nothing)
+function plot_field(field::Array{<:Complex,2}, a1::Array{<:Real,1}, a2::Array{<:Real,1}; cmap="coolwarm", vmin=nothing, vmax=nothing, label=nothing)
     # Plot real and imaginary parts with the same color scale
     if cmap == "coolwarm" && vmin == nothing && vmax == nothing
         vlim = maximum(abs.(field))
@@ -42,9 +46,9 @@ function plot_field(field::Array{<:Complex,2}, a1::Array{<:Real,1}, a2::Array{<:
     end
     figure(figsize=(4,2))
     subplot(1,2,1)
-    plot_field(real(field), a1, a2, cmap=cmap, vmin=vmin, vmax=vmax)
+    plot_field(real(field), a1, a2, cmap=cmap, vmin=vmin, vmax=vmax, label="Re[$label]")
     subplot(1,2,2)
-    plot_field(imag(field), a1, a2, cmap=cmap, vmin=vmin, vmax=vmax)
+    plot_field(imag(field), a1, a2, cmap=cmap, vmin=vmin, vmax=vmax, label="Im[$label]")
 end
 
 
@@ -78,10 +82,10 @@ end
 """
 To do
 """
-function plot_field(u::AbstractVector{<:Complex}, basis::PlaneWaveBasis; cmap="coolwarm", vmin=nothing, vmax=nothing, k0=[0,0])
+function plot_field(u::AbstractVector{<:Complex}, basis::PlaneWaveBasis; k0=[0,0], cmap="coolwarm", vmin=nothing, vmax=nothing, label=nothing)
     field = get_field(u, basis, k0=k0)
     a1, a2 = bs_to_as(basis.b1, basis.b2)
-    plot_field(field, a1, a2, cmap=cmap, vmin=vmin, vmax=vmax)
+    plot_field(field, a1, a2, cmap=cmap, vmin=vmin, vmax=vmax, label=label)
 end
 
 
@@ -92,21 +96,9 @@ Plot the permittivity and permeability of the `geometry` in real space.
 """
 function PyPlot.plot(geometry::Geometry)
     # Plot permittivity
-    plot_field(geometry.ep, geometry.a1, geometry.a2, cmap="viridis")
-    subplot(1,2,1)
[…]
