### 2020-07-04 980ce0a
docs: add thank-you note to DetecTree library
