### 2020-07-04 4ce17b6
docs: fix cross-reference to internal Zoo function
