### 2020-07-04 3029ee9
docs: fix variable name in how-to guide
