### 2020-07-04 90036c0
docs: fix formatting of note-block
