+++ docs/src/reference/index.md
@@ -9,7 +9,8 @@ Private = false
 
 ## Private interface
 
-!!! note "The `Peacock.jl` namespace" The following objects are used internally, and aren't exported to the global namespace. However, they may be useful for more advanced users, in which case you should use `Peacock.[name]` to access these.
+!!! note "The `Peacock.jl` namespace"
+    The following objects are used internally, and aren't exported to the global namespace. However, they may be useful for more advanced users, in which case you should use `Peacock.[name]` to access these.
 
 ```@autodocs
 Modules = [Peacock]
