### 2020-07-14 #2239 @dpsanders
👋 @sp94: Sorry for the delay here. I think it would be good to include some simple code in the paper showing how to produce one of the figures. You can use one of the examples from the readme.

Please have a look at the [new submission requirements](https://joss.readthedocs.io/en/latest/submitting.html). In particular, there is now a requirement for an explicit "statement of need".

In the paper it says that the package is to "analyse and predict the behaviour". Can you expand there on exactly what the package is able to calculate?

Thanks!
