### 2020-07-14 #2239 @dpsanders
Also test coverage seems rather low. I suggest you use `@testset`s. You can use an automated service like `coveralls.io` to check test coverage.
