### 2020-07-22 5b8bb46
paper: divide into summary/statement of need/examples
