### 2020-07-22 7b46dce
paper: make purpose of software clearer in statement of need
