+++ paper/paper.md
@@ -60,6 +60,15 @@ backscattering [@rider2019perspective].
 
 # Statement of need
 
+![Example reproduction of a fragilely topological photonic crystal
+[@blanco2019engineering]. (a) Unit cell of the photonic crystal.
+(b-d) Out-of-plane component of the electric field of the first three
+transverse-magnetic polarised modes at $\Gamma$. (e) Band diagram of the
+transverse-magnetic polarised modes. (f) The Wilson loop spectrum of bands 2-3
+wind, indicating non-trivial band topology. (g) The Wilson loop spectrum of the
+full valence band space does not wind, indicating that bands 2-3 are 'fragilely'
+topological. \label{fig:examples}](figures/examples.pdf)
+
 `Peacock.jl` provides a user-friendly interface to calculate and analyse
 the eigensolutions of 2D photonic crystals,
 with support for non-orthogonal unit cells and inhomogeneous permittivity and/or
@@ -73,20 +82,173 @@ and to study topology in electronic bands [@gresch2017z2pack], to our knowledge
 `Peacock.jl` is the first open-source package for studying the
 band topology in photonic crystals.
 
-# Examples
+
+# Example usage
+
+## Installation
+
+Before using `Peacock.jl` for the first time, you should install it using the built-in Julia package manager.
+```julia
+using Pkg
+Pkg.add("Peacock")
+```
+
+After installation, the `Peacock.jl` can be loaded in Julia.
+```julia
+using Peacock
+```
+
+We'll also install and load `PyPlot` to control our figures.
+```julia
+using Pkg
+Pkg.add("PyPlot")
+using PyPlot
+```
+
+
+## Defining a photonic crystal
+
+In this example we will reproduce a photonic crystal from chapter 5 of
+[@joannopoulos2008molding]. It consists of dielectric cylinders
+($\epsilon_\mathrm{cyl}=8.9, \mu_\mathrm{cyl}=1$) in air
+($\epsilon_\mathrm{air}=1, \mu_\mathrm{air}=1$). The cylinders are arranged
+on a square lattice with separation $a$, and each cylinder has a radius of $r=a/5$.
+
+First, we must define the functions `epf(x,y)` and `muf(x,y)`, which return the
+permittivity and permeability of the unit cell at $(x,y)$, where $(0,0)$
+is the center of the unit cell. We choose to work in units of length where the
+separation between cylinders is unity, $a=1$, such the radius of each
+cylinder is 0.2.
+```julia
+# Permittivity
+function epf(x,y)
+    # equation of a circle with radius 0.2a
+    if x^2+y^2 <= 0.2^2
+        # dielectric inside the circle
+        return 8.9
+    else
+        # air outside the circle
+        return 1
+    end
+end
+
+# Permeability is unity everywhere
+function muf(x,y)
+    return 1
+end
+```
+
+Now we declare the lattice parameters. The cylinders are on a square lattice,
+so our lattice vectors are orthogonal and of equal length $a$.
+```julia
+a1 = [1, 0]  # first lattice vector
+a2 = [0, 1]  # second lattice vector
+```
+
+We must also give the resolution at which the geometry should be generated at.
+```julia
+d1 = 0.01  # resolution along first lattice vector
+d2 = 0.01  # resolution along second lattice vector
+```
+A smaller value of `d1` or `d1` will result in a higher resolution grid.
+
+Finally, we are ready to construct and visualise our `Geometry`.
+```julia
+geometry = Geometry(epf, muf, a1, a2, d1, d2)
+plot(geometry)
+```
+![](figures/example_plot_geometry.png)
+
+
+## Setting up the solver
+
+In this section we will create a `Solver` that approximates the geometry using a
[…]
