### 2020-07-22 8bcc108
paper: change example code to Wilson loop winding
