+++ paper/paper.md
@@ -60,15 +60,6 @@ backscattering [@rider2019perspective].
 
 # Statement of need
 
-![Example reproduction of a fragilely topological photonic crystal
-[@blanco2019engineering]. (a) Unit cell of the photonic crystal.
-(b-d) Out-of-plane component of the electric field of the first three
-transverse-magnetic polarised modes at $\Gamma$. (e) Band diagram of the
-transverse-magnetic polarised modes. (f) The Wilson loop spectrum of bands 2-3
-wind, indicating non-trivial band topology. (g) The Wilson loop spectrum of the
-full valence band space does not wind, indicating that bands 2-3 are 'fragilely'
-topological. \label{fig:examples}](figures/examples.pdf)
-
 `Peacock.jl` provides a user-friendly interface to calculate and analyse
 the eigensolutions of 2D photonic crystals,
 with support for non-orthogonal unit cells and inhomogeneous permittivity and/or
@@ -85,169 +76,73 @@ band topology in photonic crystals.
 
 # Example usage
 
-## Installation
+![Example reproduction of a fragilely topological photonic crystal
+[@blanco2019engineering]. (a) Unit cell of the photonic crystal.
+(b-d) Out-of-plane component of the electric field of the first three
+transverse-magnetic polarised modes at $\Gamma$. (e) Band diagram of the
+transverse-magnetic polarised modes. (f) The Wilson loop spectrum of bands 2-3
+wind, indicating non-trivial band topology. (g) The Wilson loop spectrum of the
+full valence band space does not wind, indicating that bands 2-3 are 'fragilely'
+topological. \label{fig:examples}](figures/examples.pdf)
 
-Before using `Peacock.jl` for the first time, you should install it using the built-in Julia package manager.
+## Loading from the `Peacock.Zoo` submodule
+
+In this example we reproduce results that demonstrate the photonic crystal with
+'fragile' band topology that was introduced in [@blanco2019engineering].
+
+This crystal can be loaded from the `Peacock.Zoo` submodule using `make_dePaz_frag`.
 ```julia
-using Pkg
-Pkg.add("Peacock")
-```
+using Peacock, Peacock.Zoo, Parameters
 
-After installation, the `Peacock.jl` can be loaded in Julia.
-```julia
-using Peacock
-```
+# Size of the plane-wave basis
+fourier_space_cutoff = 7
 
-We'll also install and load `PyPlot` to control our figures.
-```julia
-using Pkg
-Pkg.add("PyPlot")
-using PyPlot
-```
+# Load the fragile photonic topological insulator from Blanco de Paz et al 2018
+@unpack geometry, solver, polarisation = make_dePaz_frag(fourier_space_cutoff)
 
-
-## Defining a photonic crystal
-
-In this example we will reproduce a photonic crystal from chapter 5 of
-[@joannopoulos2008molding]. It consists of dielectric cylinders
-($\epsilon_\mathrm{cyl}=8.9, \mu_\mathrm{cyl}=1$) in air
-($\epsilon_\mathrm{air}=1, \mu_\mathrm{air}=1$). The cylinders are arranged
-on a square lattice with separation $a$, and each cylinder has a radius of $r=a/5$.
-
-First, we must define the functions `epf(x,y)` and `muf(x,y)`, which return the
-permittivity and permeability of the unit cell at $(x,y)$, where $(0,0)$
-is the center of the unit cell. We choose to work in units of length where the
-separation between cylinders is unity, $a=1$, such the radius of each
-cylinder is 0.2.
-```julia
-# Permittivity
-function epf(x,y)
-    # equation of a circle with radius 0.2a
-    if x^2+y^2 <= 0.2^2
[…]
