### 2020-07-23 88b96c1
travis: remove unused matplotlib install from Documentation stage
