### 2020-07-23 f440bb6
travis: add code coverage check with coveralls.io
