+++ .travis.yml
@@ -3,6 +3,10 @@ language: julia
 notifications:
   email: false
 
+after_success:
+  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage;
+              Coveralls.submit(Coveralls.process_folder())'
+
 jobs:
   include:
     - stage: "Documentation"
