### 2020-07-23 2907cd6
Revert "travis: remove unused matplotlib install from Documentation stage"
This reverts commit 88b96c16df1db1c77dbaefa5fe2b255ed561d359.
