+++ .travis.yml
@@ -12,6 +12,10 @@ jobs:
     - stage: "Documentation"
       julia: 1.4
       os: linux
+      addons:
+        apt:
+          packages:
+            - python3-matplotlib
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
                                                Pkg.instantiate()'
