### 2020-07-23 936bd91
travis: add linux, windows, osx tests for Julia 1.4
