### 2020-07-23 8ed1561
travis: install matplotlib on linux for both test/documentation
