### 2020-07-23 d7e3310
travis: merge coveralls for different jobs
