### 2020-07-23 ff94af5
paper: wrap longer lines of example code
