+++ paper/paper.md
@@ -97,7 +97,8 @@ using Peacock, Peacock.Zoo, Parameters
 # Size of the plane-wave basis
 fourier_space_cutoff = 7
 
-# Load the fragile photonic topological insulator (Blanco de Paz et al 2019)
+# Load the fragile photonic topological insulator
+# from paper by Blanco de Paz et al 2019
 @unpack geometry, solver, polarisation = 
         make_dePaz_frag(fourier_space_cutoff)
 
