### 2020-07-23 a0b0222
docs: fix build error, devurl was undefined
