### 2020-07-24 562d437
paper: add plotting band diagram to example
