### 2020-07-24 b0d665b
paper: add missing latex dollar signs
