### 2020-07-24 fcfac14
paper: fix header formatting by removing non-breaking space
