### 2020-07-24 ab202e6
paper: fix header formatting by removing non-breaking space
