### 2020-07-24 b16f751
paper: fix broken cross-reference
