### 2020-07-24 #2239 @sp94
Hi @dpsanders ,

Sorry for the delay. I've just generated the new PDF which has some example code, the explicit statement of need, and expands on what the package can calculate.

I've linked `coveralls.io` and will get back to you in a couple more days once I have also added the `@testsets`.

Thanks for your comments!
