### 2020-07-24 d5df505
docs: add coveralls.io badge to README
