### 2020-07-24 aebed44
docs: add stable documentation badge to README
