### 2020-08-01 4c0a955
tests: add sample_path test
