### 2020-08-01 ed6563a
tests: fix the convmat and sample_path tests
