### 2020-08-01 1174094
bug fix: remove rounding errors in sample_path, remove accidental extra label
