### 2020-08-01 91759f5
bug fix: use abstract types for constructor arguments
