### 2020-08-01 fecba4c
tests: add new Project.toml for testing environment dependencies
