+++ Project.toml
@@ -7,7 +7,6 @@ version = "0.1.0"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 FFTW = "1.2"
