### 2020-08-01 b1205b0
tests: check the known eigenvalues of Wilson loops for crystal with fragile topology
