### 2020-08-01 5936462
tests: check that Wilson loop gauge and eigenvalues agree
