### 2020-08-09 3f5cae5
tests: use @testset and add more Wilson loop testing
