+++ test/runtests.jl
@@ -5,158 +5,188 @@ using LinearAlgebra
 using Peacock.Zoo
 using Parameters
 
-# Test sample_path
-begin
-    # Generate a random path
-    ks_orig = [[0.0,0.0]]
-    for n in 1:3
-        push!(ks_orig, ks_orig[end]+rand(2))
-    end
-    labels_orig = ["k$(n)" for n in 1:length(ks_orig)]
-    # Sample it
-    dk = 0.1
-    ks, labels = Peacock.sample_path(ks_orig, labels=labels_orig, dk=dk)
-    # Check same number of ks and labels
-    @test length(ks) == length(labels)
-    # Check k-point spacing is correct
-    for (k1,k2) in zip(ks,ks[2:end])
-        @test norm(k2-k1) <= dk
-    end
-    # Check labels match expected k positions
-    for (k_orig,label_orig) in zip(ks_orig,labels_orig)
-        n = findfirst([k==k_orig for k in ks])
-        @test labels[n] == label_orig
-    end
-end
 
+@testset "Band diagram tests" begin
 
-# Test convmat
-begin
-    # Create a test basis of plane waves
-    b1, b2 = [0.3,0.2], [0.5,-0.1]
-    ps, qs = Int[], Int[]
-    for p in -1:2, q in -2:3
-        push!(ps, p)
-        push!(qs, q)
-    end
-    basis = Peacock.PlaneWaveBasis(b1, b2, ps, qs)
-    # Convolution matrix of homogeneous array should be diagonal
-    x = 1 + rand(ComplexF64)
-    mat = fill(x, 10, 20)
-    cmat = Peacock.convmat(mat, basis)
-    @test all(diag(cmat) .≈ x)
-end
-
-
-# Test homogeneous systems with known solutions
-begin
-    # Create homogeneous geometry
-    ep = 1 + rand()
-    mu = 1 + rand()
-    a1 = [1,0]
-    a2 = [0,1]
-    d1 = d2 = 0.01
-    geometry = Geometry((x,y)->ep, (x,y)->mu, a1, a2, d1, d1)
-    # Check for increasing Fourier space cutoffs
-    for cutoff in [1,3,5]
-        # test with norm(b1) == norm(b2)
-        solver = Solver(geometry, cutoff)
-        for polarisation in [TE,TM]
-            modes = solve(solver, [1,0], polarisation)
-            @test isapprox(modes[1].frequency, 1/sqrt(ep*mu))
+    @testset "sample_path" begin
+        # Generate a random path
+        ks_orig = [[0.0,0.0]]
+        for n in 1:3
+            push!(ks_orig, ks_orig[end]+rand(2))
         end
-        # test with norm(b1) != norm(b2)
-        solver = Solver(geometry, cutoff, cutoff+2)
-        for polarisation in [TE,TM]
-            modes = solve(solver, [1,0], polarisation)
-            @test isapprox(modes[1].frequency, 1/sqrt(ep*mu))
+        labels_orig = ["k$(n)" for n in 1:length(ks_orig)]
+        # Sample it
+        dk = 0.1
+        ks, labels = Peacock.sample_path(ks_orig, labels=labels_orig, dk=dk)
+        # Check same number of ks and labels
+        @test length(ks) == length(labels)
+        # Check k-point spacing is correct
+        for (k1,k2) in zip(ks,ks[2:end])
+            @test norm(k2-k1) <= dk
+        end
+        # Check labels match expected k positions
+        for (k_orig,label_orig) in zip(ks_orig,labels_orig)
+            n = findfirst([k==k_orig for k in ks])
+            @test labels[n] == label_orig
         end
     end
+
 end
 
 
-# Test conversion between real space and reciprocal space lattice vectors
-a1 = rand(2)
[…]
