### 2020-08-09 #2239 @sp94
Hi @dpsanders,

I've added the test sets and according to coveralls.io I'm at 62% coverage. The remaining percentage is from the plotting functions. Do you have any recommendations in this case?

Thanks for your time
