### 2020-08-18 e81c039
docs: fix docstring Geometry
