### 2020-09-14 #2239 @dpsanders
Sorry for the delay here.

👋 Hi @crstnbr, @mdavezac and @mlxd: Would you be willing to review this submission to JOSS?
