### 2020-09-15 #2239 @sp94
Thanks for offering your time, everyone

@dpsanders I just noticed that @mdavezac and I are both at Imperial College London, so there could be a perceived conflict of interest. I don't think we've ever met or interacted before so hopefully this can be waived?
