### 2020-09-15 #2239 @kyleniemeyer
@dpsanders @sp94 Thanks for being upfront about that institutional connection—I think this is fine, though, since you have not actually worked together.
