### 2020-09-15 #2239 @danielskatz
Let me disagree - this is clearly a conflict, and we can waive it, but if there's another option where we don't have to, it would be better - that should be our first choice.
