### 2020-09-15 #2239 @mdavezac
As @sp94 said, I don't think we've interacted in any way either. If needed, I am still open to reviewing this paper.
