### 2020-09-16 #2239 @dpsanders
Thanks @mdavezac and apologies for the situation.

:wave: @datseris and @nmoran: Would you be able to review this submission to JOSS?
