### 2020-09-16 #2239 @nmoran
Hi @dpsanders, I'm happy to take a look at reviewing this work. It would be my first review for JOSS so wondering if there are some reviewing guidelines or example reviews I could look at?
