### 2020-09-16 #2239 @kyleniemeyer
@nmoran just jumping in here: we do have explicit [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), and in fact when the actual review begins (in a new GitHub issue), you will have a checklist to work through.

Thanks for your willingness to review this submission!
