### 2020-09-16 #2239 @dpsanders
Thanks @nmoran! And thanks @kyleniemeyer for the pointer.

If you browse through recent [papers published by JOSS](https://joss.theoj.org), each of them links to the open review, so you can get an idea of how they are carried out.
