### 2020-09-16 #2678 @whedon
[REVIEW]: Peacock.jl: Photonic crystals in Julia
**Submitting author:** @sp94 (<a href="http://orcid.org/0000-0003-0485-7047">Samuel John Palmer</a>)
**Repository:** <a href="https://github.com/sp94/Peacock.jl" target ="_blank">https://github.com/sp94/Peacock.jl</a>
**Version:** 0.1.1
