### 2020-09-16 30a0d14
paper: fix existing/add missing DOIs
