+++ paper/paper.bib
@@ -15,7 +15,7 @@ Photonic crystals
   pages={12576--12578},
   year={2003},
   publisher={National Acad Sciences},
-  doi={https://doi.org/10.1073/pnas.2133313100}
+  doi={10.1073/pnas.2133313100}
 }
 @article{yablonovitch1987inhibited,
   title={Inhibited spontaneous emission in solid-state physics and electronics},
@@ -26,7 +26,7 @@ Photonic crystals
   pages={2059},
   year={1987},
   publisher={APS},
-  doi={https://doi.org/10.1007/978-1-4615-1963-8_41}
+  doi={10.1007/978-1-4615-1963-8_41}
 }
 @article{john1987strong,
   title={Strong localization of photons in certain disordered dielectric superlattices},
@@ -37,7 +37,7 @@ Photonic crystals
   pages={2486},
   year={1987},
   publisher={APS},
-  doi={https://doi.org/10.1103/physrevlett.58.2486}
+  doi={10.1103/physrevlett.58.2486}
 }
 @article{joannopoulos2008molding,
   title={Molding the flow of light},
@@ -54,7 +54,7 @@ Photonic crystals
   pages={847--851},
   year={2003},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/nature01940}
+  doi={10.1038/nature01940}
 }
 
 
@@ -67,7 +67,8 @@ Electronic TIs
   number={14},
   pages={146802},
   year={2005},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.95.146802}
 }
 @article{kane2005quantum,
   title={Quantum spin Hall effect in graphene},
@@ -77,7 +78,8 @@ Electronic TIs
   number={22},
   pages={226801},
   year={2005},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.95.226801}
 }
 @article{bernevig2006quantum,
   title={Quantum spin Hall effect and topological phase transition in HgTe quantum wells},
@@ -88,7 +90,7 @@ Electronic TIs
   pages={1757--1761},
   year={2006},
   publisher={American Association for the Advancement of Science},
-  doi={https://doi.org/10.1126/science.1133734}
+  doi={10.1126/science.1133734}
 }
 
 
@@ -102,7 +104,7 @@ Photonic TIs
   pages={223901},
   year={2015},
   publisher={APS},
-  doi={https://doi.org/10.1103/physrevlett.114.223901}
+  doi={10.1103/physrevlett.114.223901}
 }
 @article{blanco2019engineering,
   title={Engineering fragile topology in photonic crystals: Topological quantum chemistry of light},
@@ -113,7 +115,7 @@ Photonic TIs
   pages={032005},
   year={2019},
   publisher={APS},
-  doi={https://doi.org/10.1103/physrevresearch.1.032005}
+  doi={10.1103/physrevresearch.1.032005}
 }
 @article{blanco2020tutorial,
   title={Tutorial: computing topological invariants in 2D photonic crystals},
@@ -124,7 +126,7 @@ Photonic TIs
   pages={1900117},
   year={2020},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/qute.201900117}
+  doi={10.1002/qute.201900117}
 }
 @article{rider2019perspective,
   title={A perspective on topological nanophotonics: Current status and future challenges},
@@ -135,7 +137,7 @@ Photonic TIs
   pages={120901},
   year={2019},
   publisher={AIP Publishing LLC},
-  doi={https://doi.org/10.1063/1.5086433}
+  doi={10.1063/1.5086433}
 }
 @article{wang2019band,
[…]
