### 2020-09-16 b18c148
paper: add another reference to work using Wilson loops
