+++ paper/paper.bib
@@ -150,6 +150,14 @@ Photonic TIs
   publisher={IOP Publishing},
   doi={10.1088/1367-2630/ab3f71}
 }
+@misc{palmer2020berry,
+    title={Berry bands and pseudo-spin of topological photonic phases},
+    author={Samuel J. Palmer and Vincenzo Giannini},
+    year={2020},
+    eprint={2009.07033},
+    archivePrefix={arXiv},
+    primaryClass={physics.optics}
+}
 
 Other software
 @article{johnson2001block,
