### 2020-09-16 9b1d2f1
paper: fix wrong separator between citations
