+++ paper/paper.md
@@ -66,7 +66,7 @@ with support for non-orthogonal unit cells and inhomogeneous permittivity and/or
 permeability. As well as the common tools for eigenmode analysis,
 such as visualising the eigenmodes or the plotting the eigenvalues on band diagrams as
 in \autoref{fig:examples}a-e, `Peacock.jl` also includes built-in using Wilson loop
-methods [@blanco2020tutorial,@palmer2020berry] to study the topology of photonic bands,
+methods [@blanco2020tutorial; @palmer2020berry] to study the topology of photonic bands,
 as in \autoref{fig:examples}f-g.
 Although there exists open-source software to study photonic bands [@johnson2001block]
 and to study topology in electronic bands [@gresch2017z2pack], to our knowledge
