### 2020-09-16 #2678 @nmoran
Myself and @mlxd are current colleagues at the same centre (ICHEC). I did not see anything specific mentioned in the COI policy that would make this a conflict of interest but thought I would check just in case?
