### 2020-09-17 92b4ecf
create Python environment
adding matplotlib to the Python environment so that repo2docker will install
