+++ environment.yml
@@ -1,0 +1,2 @@
+dependencies:
+  - matplotlib
