### 2020-09-19 ae075ee
remove assertion from main code that is covered in tests
