### 2020-09-19 1c1b9a1
use new transform function inside wilson_matrix
