### 2020-09-19 d05c069
remove obsolete ‘closed’ keyword from wilson loops
