### 2020-09-19 767ccf9
tests: remove obsolete ‘closed’ argument to wilson loop
