### 2020-09-19 87b49ff
fix transform for Hilbert spaces
