### 2020-09-19 97d0a1a
remove all references to obsolete ‘closed’ keyword
