### 2020-09-22 sp94/Peacock.jl#9 @nmoran
Markersize not defined in plot_wilson_loop_winding function
On line 172 of `wilson_loops.jl` `markersize` I get an error that markersize is not defined.

```
plot_band_diagram(my_solve, ks, dk=dk_outer, labels=labels, show_vlines=false, markersize=markersize)
```

```ERROR: LoadError: UndefVarError: markersize not defined
Stacktrace:
 [1] plot_wilson_loop_winding(::Solver, ::Array{BrillouinZoneCoordinate,1}, ::Peacock.Polarisation, ::UnitRange{Int64}; dk_outer::Float64, dk_inner::Nothing, labels::Array{Any,1}, delta_brillouin_zone::BrillouinZoneCoordinate) at ${HOME}/.julia/dev/Peacock/src/wilson_loops.jl:172
```
