### 2020-09-22 #2678 @nmoran
Installed the package alright and ran initial examples but getting the following error when trying to plot the Wilson loop winding [https://github.com/sp94/Peacock.jl/issues/9](https://github.com/sp94/Peacock.jl/issues/9).
