### 2020-09-22 be851cc
bug fix: markersize was undefined in plot_wilson_loop_winding
