+++ src/wilson_loops.jl
@@ -142,12 +142,14 @@ Perform a series of Wilson loops along `ks`, and plot the spectra on a band diag
 
 Keyword arguments
 - `dk_outer=nothing`: maximum distance between each loop (resolution of the scan)
-- `dk_inner`: maximum distance between points along a loop (resolution of the loop)
-- `labels=[]`: overwrite the labels for each `k` in `ks`
+- `dk_inner=nothing`: maximum distance between points along a loop (resolution of the loop)
 - `delta_brillouin_zone=BrillouinZoneCoordinate(0,1)`: each Wilson loop starts at and finishes in at the same `k` in different Brillouin zones
+- `labels=[]`: overwrite the labels for each `k` in `ks`
+- `markersize=nothing`: overwrite the size of each point
 """
 function plot_wilson_loop_winding(solver::Solver, ks, polarisation, bands::AbstractVector{<:Int};
-                dk_outer=nothing, dk_inner=nothing, labels=[], delta_brillouin_zone=BrillouinZoneCoordinate(0,1))
+                dk_outer=nothing, dk_inner=nothing, delta_brillouin_zone=BrillouinZoneCoordinate(0,1),
+                labels=[], markersize=nothing)
     # Convert BrillouinZoneCoordinate to labelled positions in k space
     if labels == []
         labels = [hasproperty(x,:label) ? x.label : "" for x in ks]
