### 2020-09-22 f6d45f4
add default markersize=6
markersize=nothing was causing errors on some platforms
