### 2020-09-22 e1508fb
revert to markersize=nothing and try newer matplotlib on travis instead
