### 2020-09-22 0e029a9
travis: use --user flag for pip install
