### 2020-09-22 aa03abd
travis: install matplotlib via addons
