### 2020-09-22 bb21a99
travis: try using conda to install matplotlib
