### 2020-09-23 190008c
Fix markersize=nothing undefined, run plots in tests (#10)
* bug fix: markersize was undefined in plot_wilson_loop_winding

* add plots to runtests

doesn't test outputs, but should detect basic runtime errors

* travis: install a newer version of matplotlib
