### 2020-09-23 #2678 @sp94
Thanks! sp94/Peacock.jl#9 should be fixed now. I haven't tagged a new release yet so could you run `]dev Peacock` in the REPL to check?

I missed this because I wasn't running the plot functions on Travis as I didn't know a good way to test the outputs. I've now added the plot functions to at least pick up runtime errors like this, even if the outputs are still going untested
