### 2020-09-27 #2678 @dpsanders
@PhilipVinc: Would you be able to review this submission for JOSS?
