### 2020-09-28 sp94/Peacock.jl#11 @mlxd
Examples have run errors with current tag
Hi @sp94 I am currently running the examples as in the docs, and as @nmoran pointed out the wilson loop functions not being exposed causes them to fail with the current tag.

Also, using both the tagged and dev version causes https://sp94.github.io/Peacock.jl/stable/how-tos/wilson_loops/#Plotting-the-Wilson-loop-winding to fail due to an undefined `figure` (should be fixable by adding PyPlot to the imports in that example).
