### 2020-09-28 sp94/Peacock.jl#12 @mlxd
Test with Julia LTS too
Given the compat tag of the package lists Julia as 1.x compatible, it may be a good idea to test with both the LTS version of Julia, as well as a more recent (1.4/1.5) version. Otherwise, updating the compat to the lowest supported version and testing with this should be sufficient.
