### 2020-09-28 #2678 @mlxd
hi @sp94 I have added two issues https://github.com/sp94/Peacock.jl/issues/11 and https://github.com/sp94/Peacock.jl/issues/12 to finish my review. Once these are resolved, and the final version tagged for release I am happy to give this a thumbs-up.
