### 2020-09-28 de8f76a
docs: add missing PyPlot import to example
