### 2020-09-28 sp94/Peacock.jl#11 @sp94
Thanks! I've added `using PyPlot` to the docs

I will tag a new stable release after the review
