### 2020-09-28 sp94/Peacock.jl#12 @sp94
I just tried installing PyPlot on Julia 1.0.5 and had a compatability error
```julia
(v1.0) pkg> add PyPlot
  Updating registry at `~/.julia/registries/General`
  Updating git-repo `https://github.com/JuliaRegistries/General.git`
 Resolving package versions...
ERROR: Unsatisfiable requirements detected for package PyPlot [d330b81b]:
 PyPlot [d330b81b] log:
 ├─possible versions are: [2.6.1-2.6.3, 2.7.0, 2.8.0-2.8.2, 2.9.0] or uninstalled
 └─restricted to versions 2.6.0 by an explicit requirement — no versions left
```

The lowest version that works without changes is Julia v1.3. I'll update the compat tag to 1.3 and above. I think the next LTS version of Julia is relatively soon (1.6)
