+++ Project.toml
@@ -11,4 +11,4 @@ PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 [compat]
 FFTW = "1.2"
 PyPlot = "2"
-julia = "1"
+julia = "1.3"
