### 2020-09-28 5022ba2
tests: use earliest supported Julia (v1.3)
