+++ .travis.yml
@@ -1,7 +1,7 @@
 language: julia
 
 julia:
-  - 1.4
+  - 1.3
 
 os:
   - linux
