### 2020-09-28 sp94/Peacock.jl#12 @sp94
The latest tests passed on 1.3
