### 2020-09-28 sp94/Peacock.jl#11 @sp94
Works for me on the master branch now
