### 2020-09-28 #2678 @sp94
Thanks for your time @nmoran and @mlxd! I believe I've resolved these issues now.

I've tagged v0.1.1 with these fixes and will edit once it the registration is complete
