### 2020-10-12 e237fc0
rename Mode to Eigenmode and HilbertSpace to Eigenspace
