+++ docs/src/how-tos/wilson_loops.md
@@ -50,7 +50,7 @@ title("Bands 2&3")
 ```
 ![Plot some Wilson loop examples](../figures/example_wilson_loops.png)
 
-In the first figure, the Wilson loops through the Hilbert spaces of bands 2&3 wind with Chern numbers ±1, indicating some non-trivial topology. However, the second figure shows that including the (trivial) acoustic band in the Wilson loop calculation removes the topological winding, and consequently bands 2&3 are said to be 'fragilely topological'.
+In the first figure, the Wilson loops through the eigenspaces of bands 2&3 wind with Chern numbers ±1, indicating some non-trivial topology. However, the second figure shows that including the (trivial) acoustic band in the Wilson loop calculation removes the topological winding, and consequently bands 2&3 are said to be 'fragilely topological'.
 
 
 ## Further reading
