### 2020-10-12 169a3f0
tests: shifting mode by k0 with new transforms
