### 2020-10-12 b79593e
ensure solutions are normalised
