### 2020-10-12 61e3140
rename Mode and HilbertSpace to Eigenmode and Eigenspace
