+++ src/transforms.jl
@@ -31,14 +31,14 @@ function transform(data, basis, k_map)
     return data_new
 end
 
-function transform(mode::Mode, k_map)
+function transform(mode::Eigenmode, k_map)
     k0_ = k_map(mode.k0)
     data_ = transform(mode.data, mode.basis, k_map)
-    return Mode(k0_, mode.frequency, data_, mode.weighting, mode.basis, mode.label)
+    return Eigenmode(k0_, mode.frequency, data_, mode.weighting, mode.basis, mode.label)
 end
 
-function transform(space::HilbertSpace, k_map)
+function transform(space::Eigenspace, k_map)
     k0_ = k_map(space.k0)
     data_ = transform(space.data, space.basis, k_map)
-    HilbertSpace(k0_, data_, space.weighting, space.basis)
+    Eigenspace(k0_, data_, space.weighting, space.basis)
 end
