### 2020-10-12 40e2f9f
replace obsolete shift_k0 function
