### 2020-10-12 64b2157
wilson loops: replace closed keyword with k_map keyword
