### 2020-10-12 99061c2
allow Eigenspace to be constructed from single Eigenmode
