### 2020-10-12 6e31e91
getting eigenmodes of symmetry operations
