### 2020-10-13 f291af1
allow construction of Eigenspace for a single mode
