### 2020-10-13 0e3e8ae
add symmetry_transform, symmetry_eigvals, etc, for k that are invariant (up to reciprocal lattice shift) under k_map
