### 2020-10-13 4642223
tests: symmetry eigenvalues for known example
