### 2020-10-13 1fd866b
eigenmodes: rename label to data_label, rename frequency to eigenvalue, and add eigenvalue_label
