+++ src/transforms.jl
@@ -34,13 +34,13 @@ end
 function transform(mode::Eigenmode, k_map::Function)
     k0_ = k_map(mode.k0)
     data_ = transform(mode.data, mode.basis, k_map)
-    return Eigenmode(k0_, mode.frequency, data_, mode.weighting, mode.basis, mode.label)
+    return Eigenmode(k0_, data_, mode.weighting, mode.basis, mode.eigenvalue, mode.data_label, mode.eigenvalue_label)
 end
 
 function transform(space::Eigenspace, k_map::Function)
     k0_ = k_map(space.k0)
     data_ = transform(space.data, space.basis, k_map)
-    Eigenspace(k0_, data_, space.weighting, space.basis)
+    Eigenspace(k0_, data_, space.weighting, space.basis, space.data_label)
 end
 
 
@@ -59,14 +59,14 @@ function symmetry_eigvals(space::Eigenspace, k_map::Function)
     return vals
 end
 
-function symmetry_eigen(space::Eigenspace, k_map::Function)
+function symmetry_eigen(space::Eigenspace, k_map::Function, eigenvalue_label="")
     space_ = symmetry_transform(space, k_map)
     xlaps = overlaps(space, space_)
     vals, vecs = eigen(xlaps, sortby=angle)
     modes = Eigenmode[]
     for n in 1:length(vals)
         data = space.data * vecs[:,n]
-        mode = Eigenmode(space.k0, vals[n], data, space.weighting, space.basis, "")
+        mode = Eigenmode(space.k0, data, space.weighting, space.basis, vals[n], space.data_label, eigenvalue_label)
         push!(modes, mode)
     end
     return modes
