### 2020-10-15 2419065
keep band diagram x-axis from 0 to 1
