### 2020-10-15 c700056
keep band diagram x-axis from 0 to 1
