### 2020-10-17 #2678 @sp94
Hi @dpsanders, as the reviewers are happy I've [deposited v0.1.1 to Zenodo](https://zenodo.org/record/4097237), the DOI is 10.5281/zenodo.4097236. I've set the Zenodo title and authors to match the JOSS paper.
