### 2020-10-17 ed22552
return berry curvature using new eigenmode struct
