### 2020-10-18 2fd012b
paper: clarifications for JOSS
