+++ paper/paper.bib
@@ -1,8 +1,9 @@
 PWEM
-@article{rumpf2006design,
-  title={Design and optimization of nano-optical elements by coupling fabrication to optical behavior},
-  author={Rumpf, Raymond},
-  year={2006}
+@phdthesis{rumpf2006design,
+	title={Design and optimization of nano-optical elements by coupling fabrication to optical behavior},
+	author={Rumpf, Raymond},
+	year={2006},
+	school={University of Central Florida}
 }
 
 Photonic crystals
@@ -20,7 +21,7 @@ Photonic crystals
 @article{yablonovitch1987inhibited,
   title={Inhibited spontaneous emission in solid-state physics and electronics},
   author={Yablonovitch, Eli},
-  journal={Physical review letters},
+  journal={Physical Review Letters},
   volume={58},
   number={20},
   pages={2059},
@@ -31,7 +32,7 @@ Photonic crystals
 @article{john1987strong,
   title={Strong localization of photons in certain disordered dielectric superlattices},
   author={John, Sajeev},
-  journal={Physical review letters},
+  journal={Physical Review Letters},
   volume={58},
   number={23},
   pages={2486},
@@ -39,11 +40,12 @@ Photonic crystals
   publisher={APS},
   doi={10.1103/physrevlett.58.2486}
 }
-@article{joannopoulos2008molding,
-  title={Molding the flow of light},
+@book{joannopoulos2008molding,
+  title={Photonic Crystals: Molding the flow of light},
   author={Joannopoulos, John D and Johnson, Steven G and Winn, Joshua N and Meade, Robert D},
-  journal={Princeton Univ. Press, Princeton, NJ},
-  year={2008}
+  publisher={Princeton University Press, Princeton, NJ},
+  year={2008},
+  edition={2},
 }
 @article{knight2003photonic,
   title={Photonic crystal fibres},
@@ -62,7 +64,7 @@ Electronic TIs
 @article{kane2005z,
   title={$Z_{2}$ topological order and the quantum spin Hall effect},
   author={Kane, Charles L and Mele, Eugene J},
-  journal={Physical review letters},
+  journal={Physical Review Letters},
   volume={95},
   number={14},
   pages={146802},
@@ -73,7 +75,7 @@ Electronic TIs
 @article{kane2005quantum,
   title={Quantum spin Hall effect in graphene},
   author={Kane, Charles L and Mele, Eugene J},
-  journal={Physical review letters},
+  journal={Physical Review Letters},
   volume={95},
   number={22},
   pages={226801},
@@ -82,7 +84,7 @@ Electronic TIs
   doi={10.1103/PhysRevLett.95.226801}
 }
 @article{bernevig2006quantum,
-  title={Quantum spin Hall effect and topological phase transition in HgTe quantum wells},
+  title={Quantum spin {H}all effect and topological phase transition in {H}g{T}e quantum wells},
   author={Bernevig, B Andrei and Hughes, Taylor L and Zhang, Shou-Cheng},
   journal={Science},
   volume={314},
@@ -98,7 +100,7 @@ Photonic TIs
 @article{wu2015scheme,
   title={Scheme for achieving a topological photonic crystal by using dielectric material},
   author={Wu, Long-Hua and Hu, Xiao},
-  journal={Physical review letters},
+  journal={Physical Review Letters},
   volume={114},
   number={22},
   pages={223901},
@@ -163,7 +165,7 @@ Other software
 @article{johnson2001block,
   title={Block-iterative frequency-domain methods for Maxwell's equations in a planewave basis},
   author={Johnson, Steven G and Joannopoulos, John D},
-  journal={Optics express},
+  journal={Optics Express},
   volume={8},
   number={3},
   pages={173--190},
