### 2020-10-18 90e2818
paper: remove accidental space in word
