### 2020-10-18 9237f34
paper: fix capitalisation of reference
