+++ paper/paper.bib
@@ -41,7 +41,7 @@ Photonic crystals
   doi={10.1103/physrevlett.58.2486}
 }
 @book{joannopoulos2008molding,
-  title={Photonic Crystals: Molding the flow of light},
+  title={Photonic {C}rystals: Molding the flow of light},
   author={Joannopoulos, John D and Johnson, Steven G and Winn, Joshua N and Meade, Robert D},
   publisher={Princeton University Press, Princeton, NJ},
   year={2008},
@@ -62,7 +62,7 @@ Photonic crystals
 
 Electronic TIs
 @article{kane2005z,
-  title={$Z_{2}$ topological order and the quantum spin Hall effect},
+  title={$Z_{2}$ topological order and the quantum spin {H}all effect},
   author={Kane, Charles L and Mele, Eugene J},
   journal={Physical Review Letters},
   volume={95},
@@ -73,7 +73,7 @@ Electronic TIs
   doi={10.1103/PhysRevLett.95.146802}
 }
 @article{kane2005quantum,
-  title={Quantum spin Hall effect in graphene},
+  title={Quantum spin {H}all effect in graphene},
   author={Kane, Charles L and Mele, Eugene J},
   journal={Physical Review Letters},
   volume={95},
@@ -163,7 +163,7 @@ Photonic TIs
 
 Other software
 @article{johnson2001block,
-  title={Block-iterative frequency-domain methods for Maxwell's equations in a planewave basis},
+  title={Block-iterative frequency-domain methods for {M}axwell's equations in a planewave basis},
   author={Johnson, Steven G and Joannopoulos, John D},
   journal={Optics Express},
   volume={8},
@@ -174,7 +174,7 @@ Other software
   doi={10.1364/OE.8.000173}
 }
 @article{gresch2017z2pack,
-  title={Z2Pack: Numerical implementation of hybrid Wannier centers for identifying topological materials},
+  title={Z2Pack: Numerical implementation of hybrid {W}annier centers for identifying topological materials},
   author={Gresch, Dominik and Autes, Gabriel and Yazyev, Oleg V and Troyer, Matthias and Vanderbilt, David and Bernevig, B Andrei and Soluyanov, Alexey A},
   journal={Physical Review B},
   volume={95},
