### 2020-10-18 3497c96
paper: make capitalisation of DOIs consistent
