+++ paper/paper.bib
@@ -70,7 +70,7 @@ Electronic TIs
   pages={146802},
   year={2005},
   publisher={APS},
-  doi={10.1103/PhysRevLett.95.146802}
+  doi={10.1103/physrevlett.95.146802}
 }
 @article{kane2005quantum,
   title={Quantum spin {H}all effect in graphene},
@@ -81,7 +81,7 @@ Electronic TIs
   pages={226801},
   year={2005},
   publisher={APS},
-  doi={10.1103/PhysRevLett.95.226801}
+  doi={10.1103/physrevlett.95.226801}
 }
 @article{bernevig2006quantum,
   title={Quantum spin {H}all effect and topological phase transition in {H}g{T}e quantum wells},
@@ -171,7 +171,7 @@ Other software
   pages={173--190},
   year={2001},
   publisher={Optical Society of America},
-  doi={10.1364/OE.8.000173}
+  doi={10.1364/oe.8.000173}
 }
 @article{gresch2017z2pack,
   title={Z2Pack: Numerical implementation of hybrid {W}annier centers for identifying topological materials},
@@ -182,5 +182,5 @@ Other software
   pages={075146},
   year={2017},
   publisher={APS},
-  doi={10.1103/PhysRevB.95.075146}
+  doi={10.1103/physrevb.95.075146}
 }
