### 2020-10-18 356b5a6
paper: use DOI capitalisation as given by each journal
