### 2020-10-26 #2678 @kyleniemeyer
Hi @sp94, in processing your submission, I noticed that Figure 1 references the Zi et al. 2003 PNAS paper for the copyright for part of it. In general, one would need to obtain permission from a publisher for reusing a figure from a published article, unless shared under a CC BY (or similar) license which explicitly gives permission. According to [their policies](https://www.pnas.org/page/about/rights-permissions), PNAS retains copyright for articles published up to 2008 (since the, the authors retain copyright).

Can you please reach out and get permission to use the image from that article, or find an alternative?
