### 2020-10-26 #2678 @sp94
Hi @kyleniemeyer, we contacted PNAS and they confirmed that publishing in the JOSS counts as non-commercial usage, and express permission isn't required so long as we include that copyright notice.

Here is their full reply:

> Thank you for checking. Yes, this qualifies as non-commercial use. Express permission from PNAS is not required if all of the following conditions hold true:
>
> - The PNAS article published in 1993 to present under our standard exclusive license (we started offering two other licenses in Fall 2017)
>
> - The requested material (whether figure, table, text extract) is original to the PNAS article (it does not contain previously published material, nor was adapted from an earlier source)
>
> - The intended use is for noncommercial or educational purposes
>
> If the article published between 1993 and 2008: Please include a complete citation for the original PNAS article when reusing the material, and include "Copyright (YEAR) National Academy of Sciences, U.S.A." as a copyright note.
>
> If the article published from 2009 to present: Please include a complete citation for the original PNAS article when reusing the material (do not include a copyright note).
