### 2020-10-26 #2678 @kyleniemeyer
Many thanks to @mlxd and @nmoran for reviewing this, and @dpsanders for editing it.
