### 2020-10-26 #2678 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
